# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5fee94e5a3e0491123239459
  title: Educom
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /teacher:
    post:
      x-skaffolder-id: 5ff0ad42a3e0491123239874
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ff0a150a3e049112323980c
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Teacher
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: object
                administration:
                  type: object
                contact:
                  type: object
                dob:
                  type: integer
                email:
                  type: string
                firstName:
                  type: string
                gender:
                  type: string
                lastName:
                  type: string
                nin:
                  type: string
    get:
      x-skaffolder-id: 5ff0ac5ea3e0491123239867
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ff0a150a3e049112323980c
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Teacher
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 5ff0ac5ea3e0491123239868
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ff0a150a3e049112323980c
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ff0ad42a3e0491123239875
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ff0a150a3e049112323980c
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    post:
      x-skaffolder-id: 5ff0ad42a3e0491123239873
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ff0a150a3e049112323980c
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teacher"
  "/teacher/findBysubjectId/{key}":
    get:
      x-skaffolder-id: 5ff0aea4a3e0491123239892
      x-skaffolder-name: findBysubjectId
      x-skaffolder-id-resource: 5ff0a150a3e049112323980c
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: findBysubjectId
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBysubjectId
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBysubjectId/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBysubjectId
      tags:
        - Teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Teacher:
      x-skaffolder-id: 5ff0a150a3e049112323980c
      x-skaffolder-id-db: 5fee94e5a3e049112323945a
      x-skaffolder-id-entity: 5ff0a150a3e0491123239801
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        address:
          type: object
          x-skaffolder-id-attr: 5ff0a150a3e0491123239806
          x-skaffolder-type: Custom
        administration:
          type: object
          x-skaffolder-id-attr: 5ff0a150a3e0491123239809
          x-skaffolder-type: Custom
        contact:
          type: object
          x-skaffolder-id-attr: 5ff0a150a3e0491123239807
          x-skaffolder-type: Custom
        dob:
          type: integer
          x-skaffolder-id-attr: 5ff0a150a3e0491123239804
          x-skaffolder-type: Date
        email:
          type: string
          x-skaffolder-id-attr: 5ff0a650a3e0491123239847
          x-skaffolder-type: String
        firstName:
          type: string
          x-skaffolder-id-attr: 5ff0a150a3e0491123239802
          x-skaffolder-type: String
        gender:
          type: string
          x-skaffolder-id-attr: 5ff0a150a3e0491123239808
          x-skaffolder-type: String
          x-skaffolder-enumeration:
            - FEMALE
            - MALE
        lastName:
          type: string
          x-skaffolder-id-attr: 5ff0a150a3e0491123239803
          x-skaffolder-type: String
        nin:
          type: string
          x-skaffolder-id-attr: 5ff0a150a3e0491123239805
          x-skaffolder-type: String
      x-skaffolder-relations:
        subjectId:
          x-skaffolder-id: 5ff0a9d2a3e0491123239859
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5ff0a150a3e0491123239801
          x-skaffolder-ent2: 5ff09e9ea3e04911232397fa
      required:
        - _id
...